Refactor and repackage code
