+++ src/main/java/duke/Launcher.java
@@ -4,7 +4,7 @@ import duke.ui.Main;
 import javafx.application.Application;
 
 /**
- * A launcher class to workaround classpath issues. This is the entry points to the program.
+ * A launcher class to workaround classpath issues. This is the entry point to the program.
  */
 public class Launcher {
