+++ src/main/java/duke/command/Command.java
@@ -1,8 +1,8 @@
 package duke.command;
 
 import duke.exception.DukeException;
+import duke.storage.Storage;
 import duke.task.TaskList;
-import duke.ui.Storage;
 import duke.ui.Ui;
 
 /**
