+++ src/main/java/duke/command/DeleteCommand.java
@@ -2,9 +2,9 @@ package duke.command;
 
 import duke.exception.DukeException;
 import duke.exception.InvalidIndexException;
+import duke.storage.Storage;
 import duke.task.Task;
 import duke.task.TaskList;
-import duke.ui.Storage;
 import duke.ui.Ui;
 
 /**
