+++ src/main/java/duke/command/ExitCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
 import duke.exception.StorageSavingException;
+import duke.storage.Storage;
 import duke.task.TaskList;
-import duke.ui.Storage;
 import duke.ui.Ui;
 
 /**
